% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{ag_name}
\alias{ag_name}
\title{Specify a tensor name}
\usage{
ag_name(x)
}
\arguments{
\item{x}{A string}
}
\value{
\code{x}, invisibly
}
\description{
This can be used before any autographed expression that results in the
creation of a tensor or op graph node. This can be used before \code{for} (both with tensors and datasets), \code{while}, and \code{if} statements.
}
\examples{
\dontrun{
## when you're in graph mode. (e.g, tf$executing_eagerly == FALSE)

ag_name("main-training-loop")
for(elem in dataset) ...
}
}
