% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_function_graph}
\alias{view_function_graph}
\title{Visualizes the generated graph}
\usage{
view_function_graph(fn, args, ...)
}
\arguments{
\item{args}{arguments passed to `fun`}

\item{...}{other arguments passed to [tensorflow::tensorboard()]}

\item{fun}{TensorFlow function}
}
\description{
Visualizes the generated graph
}
